tasks/zinaida/app/requirements.txt
aiohttp
aiohttp_jinja2
pycryptodome
python-barcode[images]
